refactor: add egoist.generate.clikit.runtime
