+++ egoist/data/definitions.py.tmpl
@@ -4,14 +4,12 @@ settings: SettingsDict = {"root": "cmd/", "here": __file__}
 app = App(settings)
 
 app.include("egoist.directives.define_cli")
-define_cli = app.define_cli("egoist.generate:walk")
 
 
-@define_cli
+@app.define_cli("egoist.generate.clikit:walk")
 def hello(*, name: str) -> None:
     """hello message"""
-    from egoist import runtime
-    from egoist.generate.clikit import clikit
+    from egoist.generate.clikit import runtime, clikit
 
     with runtime.generate(clikit):
         runtime.printf("hello %s\n", name)
